project_root/app/requirements.txt
streamlit==1.37.1
pillow==10.4.0
numpy==1.26.4
tensorflow==2.17.0
